Update to validation ROM flow. (#622)
