+++ rom/dev/src/flow/mod.rs
@@ -14,9 +14,12 @@ Abstract:
 
 mod cold_reset;
 mod update_reset;
+#[cfg(feature = "val-rom")]
+mod val;
 mod warm_reset;
 
-use crate::rom_env::RomEnv;
+use crate::cprintln;
+use crate::{handle_fatal_error, rom_env::RomEnv};
 use caliptra_common::FirmwareHandoffTable;
 use caliptra_drivers::{CaliptraResult, ResetReason};
 use caliptra_error::CaliptraError;
@@ -28,17 +31,35 @@ use caliptra_error::CaliptraError;
 /// * `env` - ROM Environment
 pub fn run(env: &mut RomEnv) -> CaliptraResult<Option<FirmwareHandoffTable>> {
     let reset_reason = env.soc_ifc.reset_reason();
-    match reset_reason {
-        // Cold Reset Flow
-        ResetReason::ColdReset => cold_reset::ColdResetFlow::run(env),
 
-        // Warm Reset Flow
-        ResetReason::WarmReset => warm_reset::WarmResetFlow::run(env),
+    if cfg!(not(feature = "val-rom")) {
+        match reset_reason {
+            // Cold Reset Flow
+            ResetReason::ColdReset => cold_reset::ColdResetFlow::run(env),
 
-        // Update Reset Flow
-        ResetReason::UpdateReset => update_reset::UpdateResetFlow::run(env),
+            // Warm Reset Flow
+            ResetReason::WarmReset => warm_reset::WarmResetFlow::run(env),
 
-        // Unknown/Spurious Reset Flow
-        ResetReason::Unknown => Err(CaliptraError::ROM_UNKNOWN_RESET_FLOW),
+            // Update Reset Flow
+            ResetReason::UpdateReset => update_reset::UpdateResetFlow::run(env),
+
+            // Unknown/Spurious Reset Flow
+            ResetReason::Unknown => Err(CaliptraError::ROM_UNKNOWN_RESET_FLOW),
+        }
+    } else {
+        #[allow(unused_mut)]
+        let mut result: CaliptraResult<Option<FirmwareHandoffTable>> =
+            Err(CaliptraError::ROM_GLOBAL_VAL_ROM_IN_PRODUCTION);
+
+        if env.soc_ifc.lifecycle() == caliptra_drivers::Lifecycle::Production {
+            cprintln!("Validation ROM in Production lifecycle prohibited");
+            handle_fatal_error(CaliptraError::ROM_GLOBAL_VAL_ROM_IN_PRODUCTION.into());
+        }
+        #[cfg(feature = "val-rom")]
+        {
+            result = val::ValRomFlow::run(env);
+        }
+
+        result
     }
 }
